align search filters to bottom #2942
