+++ src/Repository/UserRepository.php
@@ -229,8 +229,8 @@ class UserRepository extends ServiceEntityRepository implements PasswordUpgrader
             ->setParameter('role2', '%ROLE_ADMIN_TERRITORY%')
             ->setParameter('role3', '%ROLE_USER_PARTNER%');
 
-        if (!empty($orderType)) {
-            [$orderField, $orderDirection] = explode('-', $orderType);
+        if (!empty($searchArchivedUser->getOrderType())) {
+            [$orderField, $orderDirection] = explode('-', $searchArchivedUser->getOrderType());
             $queryBuilder->orderBy($orderField, $orderDirection);
         } else {
             $queryBuilder->orderBy('u.nom', 'ASC');
