+++ src/Service/ListFilters/SearchArchivedSignalement.php
@@ -11,6 +11,7 @@ class SearchArchivedSignalement
 
     private ?string $queryReference = null;
     private ?Territory $territory = null;
+    private ?string $orderType = null;
 
     public function getQueryReference(): ?string
     {
@@ -30,5 +31,15 @@ class SearchArchivedSignalement
     public function setTerritory(?Territory $territory): void
     {
         $this->territory = $territory;
+    }
+
+    public function getOrderType(): ?string
+    {
+        return $this->orderType;
+    }
+
+    public function setOrderType(?string $orderType): void
+    {
+        $this->orderType = $orderType;
     }
 }
