+++ templates/back/partner/index.html.twig
@@ -30,7 +30,7 @@
 
     <section class="fr-col-12">
         {{ form_errors(form) }}
-        <div class="fr-grid-row fr-grid-row--gutters">
+        <div class="fr-grid-row fr-grid-row--bottom fr-grid-row--gutters">
             <div class="fr-col-12 fr-col-lg-3">
                 {{ form_row(form.queryPartner) }}
             </div>
