+++ templates/back/signalement_archived/index.html.twig
@@ -25,7 +25,7 @@
 
     <section class="fr-container--fluid">
         {{ form_errors(form) }}
-        <div class="fr-grid-row fr-grid-row--gutters">
+        <div class="fr-grid-row fr-grid-row--bottom fr-grid-row--gutters">
             <div class="fr-col-12 fr-col-lg-4">
                 {{ form_row(form.queryReference) }}
             </div>
