+++ templates/back/tags/index.html.twig
@@ -37,7 +37,7 @@
     {{ form_start(form) }}
     <section class="fr-container--fluid">
         {{ form_errors(form) }}
-        <div class="fr-grid-row fr-grid-row--gutters">
+        <div class="fr-grid-row fr-grid-row--bottom fr-grid-row--gutters">
             <div class="fr-col-12 fr-col-lg-4">
                 {{ form_row(form.queryTag) }}
             </div>
@@ -53,7 +53,7 @@
     </section>
 
     <section class="fr-col-12 fr-py-5v">
-        <h2 class="fr-mb-0" id="desc-table">{{tags|length}} étiquette{% if tags|length > 1%}s{% endif %}</h2>
+        <h2 class="fr-mb-0" id="desc-table">{{ singular_or_plural(tags|length, 'étiquette trouvée', 'étiquettes trouvées') }}</h2>
     </section>
     
     <section class="fr-col-3 fr-pt-0">
@@ -177,10 +177,6 @@
             </div>
         </dialog>
     {% endfor %}
-
-    <section class="fr-col-12 fr-py-5v">
-        <h2 class="fr-mb-0" id="desc-table">{{ singular_or_plural(tags|length, 'étiquette trouvée', 'étiquettes trouvées') }}</h2>
-    </section>
 
     <section class="fr-col-12 fr-pt-0 fr-px-5v">
         {% set tableHead %}
