+++ templates/back/territory/index.html.twig
@@ -26,7 +26,7 @@
 
     <section class="fr-container--fluid overflow-visible">
         {{ form_errors(form) }}
-        <div class="fr-grid-row fr-grid-row--gutters">
+        <div class="fr-grid-row fr-grid-row--bottom fr-grid-row--gutters">
             <div class="fr-col-12 fr-col-lg-4">
                 {{ form_row(form.queryName) }}
             </div>
@@ -39,7 +39,7 @@
         </div>
     </section>
 
-    <section class="fr-col-12 fr-p-5v">
+    <section class="fr-col-12 fr-py-5v">
         <h2 class="fr-mb-0" id="desc-table">{{ singular_or_plural(territories|length, 'territoire trouvé', 'territoires trouvés') }}</h2>
     </section>
