+++ templates/back/user/index.html.twig
@@ -30,7 +30,7 @@
 
     <section class="fr-container--fluid overflow-visible">
         {{ form_errors(form) }}
-        <div class="fr-grid-row fr-grid-row--gutters">
+        <div class="fr-grid-row fr-grid-row--bottom fr-grid-row--gutters">
             <div class="fr-col-12 fr-col-lg-3">
                 {{ form_row(form.queryUser) }}
             </div>
@@ -54,7 +54,7 @@
             <div class="fr-col-12 fr-col-lg-3">
                 {{ form_row(form.permissionAffectation) }}
             </div>
-            <div class="fr-col-12">
+            <div class="fr-col-12 fr-col-lg-3">
                 <a href="{{ path('back_user_index') }}" class="fr-link fr-link--icon-left fr-icon-close-circle-line">Réinitialiser les résultats</a>
             </div>
         </div>
